stats: openmetrics shouldn't abuse metric { } filter strings
Once upon a time it was thought that it was a good idea to output labels for
openmetrics metrics based on the fields used in the filter { } sub-block.
For the most part, this is not useful since the labels are going to be
either (1) always the same, or (2) one of several values.

For example, consider the following metric block:

	metric foo {
		event_name = something
		...
		filter {
			abc = BAR
			def = BAZ*
		}
	}

The value of 'abc' will *always* be 'BAR' for this metric, therefore it is
useless to include it.

The value of 'def' will always start with 'BAZ' for this metric.  While it
may be tempting to justify this usage, the group_by provides a superset of
the functionality.  In other words, the following config snipped will
provide a superset of the same information.

	metric foo {
		event_name = something
		...
		filter {
			abc = BAR
			def = BAZ*
		}

		group_by = abc def
	}

Therefore, this filter { } sub-block labeling support is redundant and can
be safely removed.