stats: Rename 'name' to 'metric_name' in metric { } blocks
There was a decent amount of confusion about what what the 'name' setting
meant.  In short,

	metric foo {
		...
	}

was equivalent to:

	metric {
		name = foo
		...
	}

Unfortunately, this was often misunderstood as "the name of the *event* we
are interested in".  To avoid this confusion, we rename the setting to
'metric_name'.  So, from now on, if you want to specify the name inside the
metric block (instead of before the '{'), you must use:

	metric {
		metric_name = fo
		...
	}