lib-smtp: smtp-client-connection - Allow smtp_client_connection_connect() to be called several times.
This change allows calling smtp_client_connection_connect() several times with
different login callbacks. If the connection is already available, the
provided login callback is called almost immediately from the ioloop. This way,
several parallel asynchronous activities using the same connection can wait for
its availability and don't need to check whether it is available already. This
is mostly useful to obtain the capabilities of the server.