lib: Always use the representative when comparing event's categories
We need to be consistent about using the representative category pointer.
Otherwise, we may end up with duplicates in an event's categories array.

Since an event's categories are conceptually a set, it is reasonable to
write event using code that assumes no duplicates.  If duplicates are
present, it may lead to incorrect behavior or even crashes.