lib-smtp: smtp-client-transaction - Allow destroying connection while transaction still holds reference.
The lifetime of the transaction is controlled by the application and not so much
by the connection. The transaction holds a reference to the connection, so if
the connection is forcibly destroyed it needs, to notify the transaction of this
event.

This fixes:

Panic: file connection.c: line 938 (connection_list_deinit): assertion failed: (conn != list->connections)