lib-doveadm, doveadm: Send proxy-ttl differently in doveadm protocol
The --proxy-ttl parameter works only for mail commands. This breaks using
lib-doveadm with non-mail commands. Fix this by extending the doveadm
command "flags" parameter with "x" that specifies that the next field is
a tabescaped list of key=value pairs which can be used to specify any kind
of future command-specific options.