Update stats sample config with actually useful default
It is not clear to my whether it is intentional that you distribute default config with a non-functional stats module. 
These are some basic default metrics for (openmetrics compatible) monitoring. Still I need to print all the old metrics and try to find new equivalents. (Observation: _everyone_ on the net, in tutorials and guides use the _old metrics_ due to missing equivalent config. I haven't found _one_ new config example in a half hour search session. Looks disappointing.)

Also included `address = 127.0.0.1` in the example, either to show its usage or to provide a safer default.