lib-storage: Change storage-specific settings to regular settings roots (1/3)
This removes the need to have support for "dynamic settings".

This unfortunately also breaks Dovecot for the next two commits. The commits
are large enough that it doesn't make sense to squash them, and trying to
get them to work in the intermediate commits seems to be too much effort.